requirements.txt
# boto3==1.24.87
# botocore==1.27.87
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
fse==1.0.0
ftfy==6.1.1
gensim==4.2.0
huggingface-hub==0.11.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
langcodes==3.3.0
msgpack==1.0.4
nltk==3.7
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
wordfreq==3.0.3
